requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
groq>=0.4.0
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
lxml>=4.9.0
PyPDF2>=3.0.0
Pillow>=10.0.0
bcrypt>=4.0.0
PyJWT>=2.0.0
